refactor: extract Tauri command handlers into tauri_commands module
Move all #[tauri::command] fns out of lib.rs into tauri_commands.rs,
leaving lib.rs as pure module declarations and Tauri builder setup.
