+++ src-tauri/Cargo.toml
@@ -10,7 +10,7 @@ edition = "2024"
 # The `_lib` suffix may seem redundant but it is necessary
 # to make the lib name unique and wouldn't conflict with the bin name.
 # This seems to be only an issue on Windows, see https://github.com/rust-lang/cargo/issues/8519
-name = "starter_lib"
+name = "kodular_starter_lib"
 crate-type = ["staticlib", "cdylib", "rlib"]
 
 [build-dependencies]
